+++ src/main/java/com/ecommerce/ecommApp/commons/pojo/products/Product.java
@@ -1,11 +1,9 @@
 package com.ecommerce.ecommApp.commons.pojo.products;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
 import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 @Entity
@@ -14,7 +12,7 @@ public class Product {
 
   @JsonProperty
   @Getter @Id @Column(unique = true) @GeneratedValue(strategy = GenerationType.AUTO)
-  int productid;
+  long productid;
 
   public enum Gender {MALE, FEMALE};
 
@@ -28,6 +26,6 @@ public class Product {
 
   @JsonProperty
   @Getter @NotNull
-  int inventoryid;
+  long inventoryid;
 
 }
